Add db cleanup to each iteration
