+++ core/bench_test.go
@@ -6,7 +6,7 @@ import (
 )
 
 func Benchmark_Get(b *testing.B) {
-	_, db := SetupTempDB(b, WithMergeEnabled(false))
+	db, _, _ := SetupTempDB(b, WithMergeEnabled(false))
 
 	// preload some keys so Get has something to fetch
 	for i := 0; i < 10000; i++ {
@@ -27,7 +27,7 @@ func Benchmark_Get(b *testing.B) {
 }
 
 func Benchmark_Set(b *testing.B) {
-	_, db := SetupTempDB(b, WithMergeEnabled(false))
+	db, _, _ := SetupTempDB(b, WithMergeEnabled(false))
 
 	// Run the timed loop of b.N Set calls
 	b.ResetTimer()
@@ -40,7 +40,7 @@ func Benchmark_Set(b *testing.B) {
 }
 
 func Benchmark_Fsync_Set(b *testing.B) {
-	_, db := SetupTempDB(b, WithFsync(true), WithMergeEnabled(false))
+	db, _, _ := SetupTempDB(b, WithFsync(true), WithMergeEnabled(false))
 
 	// Run the timed loop of b.N Set calls
 	b.ResetTimer()
@@ -49,5 +49,40 @@ func Benchmark_Fsync_Set(b *testing.B) {
 		if err := db.Set(key, "value"); err != nil {
 			b.Fatalf("db.set: %v", err)
 		}
+	}
+}
+
+func Benchmark_Merge(b *testing.B) {
+	const (
+		rollover        = 1024 // 1KB segments
+		mergeThreshold  = 5    // start merge after 5 inactive segments
+		recordsPerBatch = 50   // writes per segment to exceed the threshold
+	)
+
+	for i := 0; i < b.N; i++ {
+		b.StopTimer()
+		db, _, cleanup := SetupTempDB(b,
+			WithRolloverThreshold(rollover),
+			WithMergeThreshold(mergeThreshold),
+			WithMergeEnabled(false),
+		)
+
+		for seg := 0; seg < mergeThreshold; seg++ {
+			for r := 0; r < recordsPerBatch; r++ {
+				key := fmt.Sprintf("key%03d%02d", seg, r)
+				val := fmt.Sprintf("val%03d%02d", seg, r)
+				if err := db.Set(key, val); err != nil {
+					b.Fatalf("set: %v", err)
+				}
+			}
+		}
+
+		b.StartTimer()
+		if err := db.merge(); err != nil {
+			b.Fatalf("merge: %v", err)
+		}
+
+		// cleanup the db at the end of each iteration
+		cleanup()
 	}
 }
