+++ core/db_test.go
@@ -11,7 +11,7 @@ import (
 )
 
 func TestSetAndGet(t *testing.T) {
-	_, db := SetupTempDB(t, WithMergeEnabled(false))
+	db, _, _ := SetupTempDB(t, WithMergeEnabled(false))
 
 	// set a key and retrieve it
 	if err := db.Set("foo", "bar"); err != nil {
@@ -26,7 +26,7 @@ func TestSetAndGet(t *testing.T) {
 }
 
 func TestOverwrite(t *testing.T) {
-	_, db := SetupTempDB(t, WithMergeEnabled(false))
+	db, _, _ := SetupTempDB(t, WithMergeEnabled(false))
 
 	// set a key twice
 	_ = db.Set("key", "first")
@@ -40,7 +40,7 @@ func TestOverwrite(t *testing.T) {
 }
 
 func TestKeyNotFound(t *testing.T) {
-	_, db := SetupTempDB(t, WithMergeEnabled(false))
+	db, _, _ := SetupTempDB(t, WithMergeEnabled(false))
 
 	if _, err := db.Get("missing"); !errors.Is(err, ErrKeyNotFound) {
 		t.Errorf("expected KeyNotFoundError, got %v", err)
@@ -48,7 +48,7 @@ func TestKeyNotFound(t *testing.T) {
 }
 
 func TestPersistence(t *testing.T) {
-	path, db := SetupTempDB(t, WithMergeEnabled(false))
+	db, path, _ := SetupTempDB(t, WithMergeEnabled(false))
 
 	_ = db.Set("a", "1")
 	_ = db.Set("b", "2")
@@ -70,7 +70,7 @@ func TestPersistence(t *testing.T) {
 }
 
 func TestLoadIndexOverwrite(t *testing.T) {
-	path, db := SetupTempDB(t, WithMergeEnabled(false))
+	db, path, _ := SetupTempDB(t, WithMergeEnabled(false))
 
 	_ = db.Set("foo", "first")
 	_ = db.Set("foo", "second")
@@ -85,7 +85,7 @@ func TestLoadIndexOverwrite(t *testing.T) {
 }
 
 func TestEmptyDB(t *testing.T) {
-	_, db := SetupTempDB(t, WithMergeEnabled(false))
+	db, _, _ := SetupTempDB(t, WithMergeEnabled(false))
 
 	if _, err := db.Get("nope"); !errors.Is(err, ErrKeyNotFound) {
 		t.Errorf("expected KeyNotFoundError on empty DB, got %v", err)
@@ -93,7 +93,7 @@ func TestEmptyDB(t *testing.T) {
 }
 
 func TestManyKeys(t *testing.T) {
-	_, db := SetupTempDB(t, WithMergeEnabled(false))
+	db, _, _ := SetupTempDB(t, WithMergeEnabled(false))
 
 	for i := 0; i < 1000; i++ {
 		k, v := fmt.Sprintf("k%03d", i), fmt.Sprintf("v%03d", i)
@@ -109,7 +109,7 @@ func TestManyKeys(t *testing.T) {
 }
 
 func TestTruncatedHeader(t *testing.T) {
-	dir, _ := SetupTempDB(t, WithMergeEnabled(false))
+	_, dir, _ := SetupTempDB(t, WithMergeEnabled(false))
 
 	// Manually write a valid record + only half of the next header
 	f, _ := os.Create(filepath.Join(dir, "seg001"))
@@ -134,7 +134,7 @@ func TestTruncatedHeader(t *testing.T) {
 }
 
 func TestTruncatedKey(t *testing.T) {
-	dir, _ := SetupTempDB(t, WithMergeEnabled(false))
+	_, dir, _ := SetupTempDB(t, WithMergeEnabled(false))
 
 	// write header for keyLen=3,valLen=2, then only 1 byte of the key
 	f, _ := os.Create(filepath.Join(dir, "seg001"))
@@ -154,7 +154,7 @@ func TestTruncatedKey(t *testing.T) {
 }
 
 func TestTruncatedValue(t *testing.T) {
-	dir, _ := SetupTempDB(t, WithMergeEnabled(false))
+	_, dir, _ := SetupTempDB(t, WithMergeEnabled(false))
 
 	// write one good record, then header+full key, but only 1 of 2 value bytes
 	f, _ := os.Create(filepath.Join(dir, "seg001"))
@@ -183,7 +183,7 @@ func TestTruncatedValue(t *testing.T) {
 }
 
 func TestOverwriteAfterPartialAppend(t *testing.T) {
-	dir, db := SetupTempDB(t, WithMergeEnabled(false))
+	db, dir, _ := SetupTempDB(t, WithMergeEnabled(false))
 
 	// 1) Write two good records: "a"→"1", "b"→"2"
 	_ = db.Set("a", "1")
@@ -250,7 +250,7 @@ func TestSegmentCount(t *testing.T) {
 	)
 
 	// open with tiny segment threshold
-	_, db := SetupTempDB(t, WithRolloverThreshold(int64(rolloverThreshold)), WithMergeEnabled(false))
+	db, _, _ := SetupTempDB(t, WithRolloverThreshold(int64(rolloverThreshold)), WithMergeEnabled(false))
 
 	// drive all the writes
 	for r := 0; r < rounds; r++ {
@@ -283,7 +283,7 @@ func TestSegmentCount(t *testing.T) {
 }
 
 func TestGetLatestWinsAcrossSegments(t *testing.T) {
-	_, db := SetupTempDB(t, WithRolloverThreshold(1), WithMergeEnabled(false)) // force a new segment per write
+	db, _, _ := SetupTempDB(t, WithRolloverThreshold(1), WithMergeEnabled(false)) // force a new segment per write
 
 	_ = db.Set("k", "v1")
 	_ = db.Set("k", "v2")
@@ -295,7 +295,7 @@ func TestGetLatestWinsAcrossSegments(t *testing.T) {
 }
 
 func TestRecoveryAcrossSegmentBoundary(t *testing.T) {
-	dir, db := SetupTempDB(t, WithRolloverThreshold(16), WithMergeEnabled(false))
+	db, dir, _ := SetupTempDB(t, WithRolloverThreshold(16), WithMergeEnabled(false))
 
 	// ─── SETUP: roll three segments by overwriting the same key ───
 	_ = db.Set("foo", "A")
@@ -330,7 +330,7 @@ func TestRecoveryAcrossSegmentBoundary(t *testing.T) {
 // the value from the segment that appears last in the file, regardless of its
 // numeric id.
 func TestManifestOrderingAffectsWinner(t *testing.T) {
-	dir, db := SetupTempDB(t, WithRolloverThreshold(1), WithMergeEnabled(false)) // force 1 key per segment
+	db, dir, _ := SetupTempDB(t, WithRolloverThreshold(1), WithMergeEnabled(false)) // force 1 key per segment
 
 	_ = db.Set("k", "old") // seg001
 	_ = db.Set("k", "new") // seg002 (last-writer-wins originally)
@@ -357,7 +357,7 @@ func TestManifestOrderingAffectsWinner(t *testing.T) {
 // with a new id but before any bytes were written to that file.  On reopen the
 // DB should reuse the zero-byte file as its active writer.
 func TestEmptyTailSegmentReuse(t *testing.T) {
-	dir, db := SetupTempDB(t, WithMergeEnabled(false))
+	db, dir, _ := SetupTempDB(t, WithMergeEnabled(false))
 	_ = db.Set("a", "1") // seg001 with data
 
 	// Force-create an empty seg002 and *do not* write to it.
